+++ src/option.rs
@@ -3,11 +3,11 @@ use crate::error_message::ErrorMessage;
 use crate::WithContext;
 
 impl<T> WithContext<T, Infallible> for Option<T> {
-    fn with_err_context(self, context: impl AsRef<str>) -> Result<T, ErrorMessage> {
+    fn with_err_context(self, context: impl ToString) -> Result<T, ErrorMessage> {
         match self {
             Some(value) => Ok(value),
             None => {
-                ErrorMessage::err(context.as_ref())
+                ErrorMessage::err(context)
             }
         }
     }